Eagerly upload completed container logs to S3
Moves S3 log backup logic into the fileserver sidecar.
Rather than waiting until the pod is being deleted before uploading,
the sidecar uploads the logs for each run of the main container as soon
as it detects that container's exit (via a sentinel file).

This also addresses a bug where log files are not uploaded to S3 if the
main app container is in a CrashBackOffLoop when the pod is deleted.