Fix type confusion when SPNEGO completes with no response
Previously this code was only taking a response from the chan if it had
a non-nil token from .acceptSecContext. If .acceptSecContext returns nil
without throwing (which means that negotiation is complete and no
WWW-Authenticate reply needs to be sent), we also need to ensure we
return an actual response, not a chan, or we'll confuse later code
that's accepting a map.